+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+### Added
+
+- Optional `workspaceId` on `Qualtive(...)`, sent as the `X-Workspace` header.